Warning fix for implicite cast in scaling from int64 to int. BUG=none TESTED=local visual c build R=tpsiaki@google.com
Review URL: https://webrtc-codereview.appspot.com/10169004

git-svn-id: http://libyuv.googlecode.com/svn/trunk@985 16f28f9a-4ce2-e073-06de-1de4eb20be90

+++ source/scale_argb.cc
@@ -186,7 +186,7 @@ static void ScaleARGBBilinearDown(int src_width, int src_height,
   if (xr > src_width) {
     xr = src_width;
   }
-  clip_src_width = (xr - xl) * 4;  // Width aligned to 4.
+  clip_src_width = (int)(xr - xl) * 4;  // Width aligned to 4.
   src_argb += xl * 4;
   x -= (int)(xl << 16);
 #if defined(HAS_INTERPOLATEROW_SSE2)
